LPS-119818 Move method to get field options to RulesSupport
