+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/RuleEditor/RuleEditor.es.js
@@ -30,6 +30,7 @@ import {Config} from 'metal-state';
 
 import {maxPageIndex, pageOptions} from '../../util/pageSupport.es';
 import {getFieldProperty} from '../LayoutProvider/util/fields.es';
+import RulesSupport from '../RuleBuilder/RulesSupport.es';
 import templates from './RuleEditor.soy';
 
 const fieldOptionStructure = Config.shapeOf({
@@ -170,19 +171,6 @@ class RuleEditor extends Component {
 		return promise;
 	}
 
-	getFieldOptions(fieldName) {
-		let options = [];
-		const visitor = new PagesVisitor(this.pages);
-
-		const field = visitor.findField((field) => {
-			return field.fieldName === fieldName;
-		});
-
-		options = field ? field.options : [];
-
-		return options;
-	}
-
 	getFieldsByTypes(fields, types) {
 		return fields.filter((field) =>
 			types.some((fieldType) => field.type == fieldType)
@@ -308,7 +296,10 @@ class RuleEditor extends Component {
 
 				operators = this._getOperatorsByFieldType(dataType);
 
-				firstOperandOptions = this.getFieldOptions(fieldName);
+				firstOperandOptions = RulesSupport.getFieldOptions(
+					fieldName,
+					pages
+				);
 			}
 
 			return {
